requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=base.txt base.in
#
django-filter==2.2.0      # via -r base.in
django==2.2.12            # via -r base.in, django-filter, djangorestframework
djangorestframework==3.11.0  # via -r base.in
markdown==3.2.1           # via -r base.in
psycopg2==2.8.5           # via -r base.in
pytz==2019.3              # via django
sqlparse==0.3.1           # via django

# The following packages are considered to be unsafe in a requirements file:
# setuptools
